pyproject.toml
[project]
name = "xlsclients"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-xlib>=0.33",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
